partial implemenation of human player
gui can get clicks and translate them to rows and columns. There is some
basic logic to hold clicks, rows, columns, relevent info, etc in a
Gui_Move object. There are queues set up in checkergui for sending the
move to repagcheckers and for verifying whether the move was legal, but
said queues are not yet implemented fully.